+++ azure-communication-ui/calling/src/main/java/com/azure/android/communication/ui/calling/models/CallCompositeDiagnosticsInfoExtensions.kt
@@ -1,0 +1,8 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+package com.azure.android.communication.ui.calling.models
+
+internal fun CallCompositeDiagnosticsInfo.setCallId(lastKnownCallId: String?) {
+    this.lastKnownCallId = lastKnownCallId
+}
